Custom Request object
Replacement Request object that adds a `getContentAsStream()` method. This
can be used to allow parsing the incoming request content as a stream, and thus
allow reducing memory usage.

Also adds a specialized request factory for returning this particular request
type, and registers it with the service manager to replace the standard HTTP
request object.