Implemented replacement RenameUpload filter
New filter compses a Request instance. On PUT/PATCH requests, it will
override the `moveUploadedFile()` method in order to do a simple file
rename instead of use `move_uploaded_file()`.

If no request instance is present, or for other request types or
methods, it falls back to the original logic of the ZF2 filter.