Allow dash-separated keys
Per a number of reports, some modules were using dash-separated keys instead of
underscore_separated keys for content-negotiation options.
zf-content-negotiation was ignoring these previously, but starting with 1.2.0
and the introduction of the ContentNegotiationOptions class, these are now
invalid and raise an exception, breaking applications.

The specific module in question is likely zf-apigility-doctrine, and there's a
PR against that repo to fix it: zfcampus/zf-apigility-doctrine#212.

This patch modifies the options class to munge dash-separated values to ensure
they work.