Fixes uninitialized string offset error
As reported in #96, since #94 and the 1.3.1 release, if the body content
for a PUT, PATCH, or DELETE call is empty, an "Uninitialized string
offset: 0" notice is raised.

This patch adds tests missing from the original patch created for #94,
verifying that JSON deserialization occurs for those HTTP methods if the
content looks like JSON but no Content-Type header was present. The new
cases also include scenarios where no content is present. A fix is
introduced to the `ContentTypeListener` to ensure the notice is no
longer raised.