Allow using zend-console to determine if a CLI env is in play
This patch updates #86 to fix #98 by testing if the
`Zend\Console\Console` class is present and, if so, using the return
value of its `isConsole()` method to determine the current SAPI
environment. This approach allows overriding the environment for
purposes of unit testing incoming HTTP requests.

If the class is not available, we fallback to the `PHP_SAPI` value.